+++ mono_cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Removed export of `mono_core`
 - `DefaultCommandEnvironmentBuilder` now uses injected core dependencies
 - New: `DefaultCliEngine` implements core `CliEngine`
+- `DefaultCliEngine` now uses injected router factory instead of creating router directly
 
 ## 0.0.5